+++ example/android/app/src/main/kotlin/com/example/swagger_mapper/MainActivity.kt
@@ -1,6 +1,0 @@
-package com.example.swagger_mapper
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}