+++ example/lib/swagger_generated_code/pet_service_yaml.swagger.chopper.dart
@@ -6,15 +6,16 @@ part of 'pet_service_yaml.swagger.dart';
 // ChopperGenerator
 // **************************************************************************
 
-// ignore_for_file: always_put_control_body_on_new_line, always_specify_types, prefer_const_declarations, unnecessary_brace_in_string_interps
-class _$PetServiceYaml extends PetServiceYaml {
+// coverage:ignore-file
+// ignore_for_file: type=lint
+final class _$PetServiceYaml extends PetServiceYaml {
   _$PetServiceYaml([ChopperClient? client]) {
     if (client == null) return;
     this.client = client;
   }
 
   @override
-  final definitionType = PetServiceYaml;
+  final Type definitionType = PetServiceYaml;
 
   @override
   Future<Response<dynamic>> _petPost({