+++ example/lib/swagger_generated_code/some_file_name.swagger.chopper.dart
@@ -6,15 +6,16 @@ part of 'some_file_name.swagger.dart';
 // ChopperGenerator
 // **************************************************************************
 
-// ignore_for_file: always_put_control_body_on_new_line, always_specify_types, prefer_const_declarations, unnecessary_brace_in_string_interps
-class _$SomeFileName extends SomeFileName {
+// coverage:ignore-file
+// ignore_for_file: type=lint
+final class _$SomeFileName extends SomeFileName {
   _$SomeFileName([ChopperClient? client]) {
     if (client == null) return;
     this.client = client;
   }
 
   @override
-  final definitionType = SomeFileName;
+  final Type definitionType = SomeFileName;
 
   @override
   Future<Response<dynamic>> _petPost({