+++ lib/src/swagger_models/swagger_tag.g.dart
@@ -1,0 +1,18 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'swagger_tag.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+SwaggerTag _$SwaggerTagFromJson(Map<String, dynamic> json) => SwaggerTag(
+      name: json['name'] as String? ?? '',
+      description: json['description'] as String? ?? '',
+    );
+
+Map<String, dynamic> _$SwaggerTagToJson(SwaggerTag instance) =>
+    <String, dynamic>{
+      'name': instance.name,
+      'description': instance.description,
+    };